requirements.base.txt
# Neo4j Core

neo4j==5.26

neo4j-driver==5.26

graphdatascience

py2neo==2021.2.4

neomodel



# Graph Algorithms and Path Finding

networkx==3.2.1

python-igraph==0.11.3

networkit==10.1

pyvis==0.3.2

graphviz==0.20.1



 

# LLM and Vector Database Integration

langchain==0.3.14

langchain-neo4j==0.2.0

transformers==4.36.2

sentence-transformers==2.2.2

faiss-cpu==1.7.4

chromadb==0.4.22

openai==1.6.1

tiktoken==0.5.2





# Data Science and Visualization

pandas==2.1.4

numpy==1.26.3

scikit-learn==1.3.2

matplotlib==3.8.2

seaborn==0.13.1

plotly==5.18.0

jupyter



# Development Tools

pylint

black

flake8

isort

pytest

pytest-cov



# Utilities

tqdm==4.66.1

requests==2.31.0

beautifulsoup4==4.12.2

faker==22.5.1

python-dateutil==2.8.2

python-dotenv==1.0.0